Add art direction to motto-section and fix problem with staff-section bg
The problem at staff-section was related with media queries used to define width of
the browser. It caused that a wrong picture was used for certain width of the browser.
What is interesting is a fact that the problem only appeard on retina displays.
The problem was solved after changing em's to px in media queries used inside of the
picture's <source media="">